Extract jwtCache to separate file
